+++ modules/postgres/main.tf
@@ -46,6 +46,13 @@ resource "azurerm_postgresql_flexible_server" "this" {
   depends_on = [
     azurerm_private_dns_zone_virtual_network_link.this
   ]
+
+  lifecycle {
+    ignore_changes = [
+      zone,
+      high_availability[0].standby_availability_zone
+    ]
+  }
 }
 
 resource "azurerm_postgresql_flexible_server_configuration" "azure_extensions" {