+++ versions.tf
@@ -13,5 +13,9 @@ terraform {
       source  = "hashicorp/helm"
       version = ">= 3.0.2"
     }
+    mongodbatlas = {
+      source  = "mongodb/mongodbatlas"
+      version = "~> 1.11"
+    }
   }
 }